may have gotten to the bottom of this log problem. not handling multiple processes correctly. removing for now.
